fix: Move createToken() to module utils.ts
Otherwise the code in src/index.ts gets called when __tests__/e2e.ts is
run which results in the call of `process.exit()` before the tests are
run.
